perf: replace scalar voxel expansion with unified ChunkBox AABB pipeline for billion-scale fills

+++ src/main/java/com/pixel/qve/neoforge/api/BatchWriteResult.java
@@ -20,7 +20,7 @@ public record BatchWriteResult(
         int totalSubmitted,
         int totalSucceeded,
         int totalFailed,
-        int totalBlocks,
+        long totalBlocks,
         int ramChunkCount,
         int diskChunkCount,
         long durationNanos,
